Add handling for hamming dataset having 1 label.
Particularly for hamming distance 1, it is possible for the dataset
generated by the hamming function to have only a single label. In this
case, rpart cannot train a tree on the dataset. Instead, we just store
that label, and have that "tree" predict that label for every instance
it classifies.